Address Rovo Dev session index review feedback
